Fixes a bug on LeadingUAV random_move.
We were not predicting whether or not the LeadingUAV will collide with
the floor.
We where just trying to making sure that if it collides, the next move
would remove it from the floor.
The problem is getting noticable, if the next few iterations the
random_move result to a small value on the z axis, thus the UAV remains
stuck on the ground.