Re-designs the Controller and Kalman Filter logic
- Separates the Controller, that just converts a target point on the map
to velocity for the EgoUAV, and the Filter which smooths out the
measurements and predicts position if there is none.

- Realizes the fact that since the Kalman Filter will predict the
movement of the LeadingUAV, using Newton's equations, the measured
position can not be an offset. It needs to be the actual position inside
EgoUAV's coordinate frame.

- Cleans up the mess with the two different coordinate frames that are
introduced, when rotating the EgoUAV (yaw).