Updates logger to include groud truth and estimation
The goal is to be able to display both the ground truth velocities
and positions of the two UAVs. At a later timestep, when the bbox
is available we may estimante those values, using the bbox and
append that information onto each frame.

We need to be careful onto which frame we append the bbox info.

Next:
- fix the angle estimation and ground truth values (normalizing
will probably be sufficient)
- verify on the fly whether the frame that you append the info
on is the correct one. You could do this by passing as an argument
the global frame index (watch out for the difference between the
sim_fps and camera_fps indexes) to the logger.update_frame() func.
- Add Graphing Functions to graph all this information you have
collected for each frame.