Adds mAP to each checkpoint and a function that calculates it
This function will load all test images and evaluate them in batches.
Then using the torchmetrics library it will use the predictions
and the targets (boxes, labels and scores) to calculate mAP.

The problem with the above implementation is that it requires a
separate pass, evaluating the test images, so we may extract from
the pytorch model both the losses and the predictions.
( Note: PyTorch will return losses if model is in training mode or
predictions if the model is in evaluation mode)